@@ -3,6 +3,14 @@
   "version": "1.0.0",
   "description": "JavaScript development environment Pluralsight course by Cory House",
   "scripts": {
+    "prestart": "babel-node buildScripts/startMessage.js",
+    "start": "npm-run-all --parallel open:src localtunnel lint:watch test:watch",
+    "open:src": "babel-node buildScripts/srcServer.js",
+    "lint": "esw webpack.config.* src buildScripts --color",
+    "lint:watch": "npm run lint -- --watch",
+    "localtunnel": "lt --port 3000",
+    "test": "mocha --reporter progress buildScripts/testSetup.js \"src/**/*.test.js\"",
+    "test:watch": "npm run test -- --watch"
   },
   "author": "Cory House",
   "license": "MIT",
@@ -10,38 +18,38 @@
     "whatwg-fetch": "1.0.0"
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-loader": "6.2.5",
-    "babel-preset-latest": "6.16.0",
-    "babel-register": "6.16.3",
-    "chai": "3.5.0",
-    "chalk": "1.1.3",
-    "cheerio": "0.22.0",
-    "compression": "1.6.2",
-    "cross-env": "3.1.3",
-    "css-loader": "0.25.0",
-    "eslint": "3.8.1",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-watch": "2.1.14",
-    "express": "4.14.0",
-    "extract-text-webpack-plugin": "1.0.1",
-    "html-webpack-plugin": "2.22.0",
-    "jsdom": "9.8.0",
-    "json-schema-faker": "0.3.6",
-    "json-server": "0.8.22",
-    "localtunnel": "1.8.1",
-    "mocha": "3.1.2",
-    "nock": "8.1.0",
-    "npm-run-all": "3.1.1",
-    "nsp": "2.6.2",
-    "numeral": "1.5.3",
+    "babel-cli": "~6.16.0",
+    "babel-core": "~6.17.0",
+    "babel-eslint": "~7.2.3",
+    "babel-loader": "^6.4.1",
+    "babel-preset-latest": "~6.16.0",
+    "babel-register": "~6.16.3",
+    "chai": "~3.5.0",
+    "chalk": "~1.1.3",
+    "cheerio": "~0.22.0",
+    "compression": "^1.7.3",
+    "cross-env": "^3.2.4",
+    "css-loader": "~0.25.0",
+    "eslint": "~3.8.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-watch": "~2.1.14",
+    "express": "^4.16.4",
+    "extract-text-webpack-plugin": "~1.0.1",
+    "html-webpack-plugin": "~2.22.0",
+    "jsdom": "^9.12.0",
+    "json-schema-faker": "^0.4.7",
+    "json-server": "^0.14.0",
+    "localtunnel": "^1.9.1",
+    "mocha": "^5.2.0",
+    "nock": "^10.0.1",
+    "npm-run-all": "^3.1.2",
+    "numeral": "^1.5.6",
     "open": "0.0.5",
-    "rimraf": "2.5.4",
-    "style-loader": "0.13.1",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.0",
-    "webpack-md5-hash": "0.0.5"
+    "rimraf": "~2.5.4",
+    "style-loader": "^0.13.2",
+    "webpack": "^1.15.0",
+    "webpack-dev-middleware": "~1.8.4",
+    "webpack-hot-middleware": "^2.24.3",
+    "webpack-md5-hash": "0.0.6"
   }
 }
